Change the expected exception for all BigOutputFileName tests, from ProcessException to ServerErrorException.
Create a single helper method in order to easily change in the future.

Details:
A server error (500) occurs when a file with a name length of 130 characters is sent. The server was expected to respond with a ProcessException error but it responded with an unspecified error (500).
{
    "error":{
       "type":"ServerError",
       "message":"Something on our end went wrong, probably we are not catching some exception we should catch! We are logging this and we will fix it.",
       "code":"500"
    }
}

System.Net.Http.HttpRequestException: 500 (Internal Server Error)